requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==23.1.0
black==23.7.0
charset-normalizer==3.2.0
click==8.1.7
Django==4.2.5
django-after-response==0.2.2
django-environ==0.10.0
django-sendgrid-v5==1.2.2
djangorestframework==3.14.0
environ==1.0
frozenlist==1.4.0
gunicorn==20.1.0
idna==3.4
multidict==6.0.4
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.1
Pillow==10.0.0
platformdirs==3.9.1
python-http-client==3.3.7
pytz==2023.3.post1
sendgrid==6.10.0
sqlparse==0.4.3
starkbank-ecdsa==2.2.0
tomli==2.0.1
typing_extensions==4.7.1
whitenoise==6.2.0
yarl==1.9.2
